+++ Net_Core/Web.Angular/projects/gw-frontend/src/app/skins/arc/arc-footer/arc-footer.component.ts
@@ -2,6 +2,7 @@ import { Component } from '@angular/core';
 
 @Component({
 	selector: 'gw-frontend-arc-footer',
+	standalone: true,
 	templateUrl: './arc-footer.component.html',
 	styleUrls: ['./arc-footer.component.scss']
 })